add auth

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,12 @@ use App\Post;
 
 class PostController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index', 'show']);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -40,6 +46,7 @@ class PostController extends Controller
         $post = new Post;
         $post->title = $request->title;
         $post->body = $request->body;
+        $post->user_id = $request->user()->id;
         $post->save();
         return redirect('posts/'.$post->id);
     }
@@ -63,6 +70,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('edit', $post);
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -75,6 +83,7 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        $this->authorize('edit', $post);
         $post->title = $request->title;
         $post->body = $request->body;
         $post->save();
@@ -89,6 +98,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('edit', $post);
         $post->delete();
         return redirect('posts');
     }
