+++ app/Http/Controllers/UserController.php
@@ -7,6 +7,12 @@ use App\User;
 
 class UserController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index', 'show']);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -65,6 +71,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize('edit', $user);
         return view('users.edit', ['user' => $user]);
     }
 
@@ -77,6 +84,7 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        $this->authorize('edit', $user);
         $user->name = $request->name;
         $user->save();
         return redirect('users/'.$user->id);
@@ -90,6 +98,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('edit', $user);
         $user->delete();
         return redirect('users');
     }
